+++ src/client/request/graphql/generators/credit-card-tokenization.js
@@ -17,6 +17,7 @@ function createMutation(config) {
     '      bin ' +
     '      brandCode ' +
     '      last4 ' +
+    '      cardholderName ' +
     '      expirationMonth' +
     '      expirationYear' +
     '      binData { ' +