+++ src/hosted-fields/internal/components/index.js
@@ -1,6 +1,7 @@
 'use strict';
 
 module.exports = {
+  cardholderName: require('./cardholder-name-input').CardholderNameInput,
   number: require('./credit-card-input').CreditCardInput,
   expirationDate: require('./expiration-date-input').ExpirationDateInput,
   expirationMonth: require('./expiration-month-input').ExpirationMonthInput,